+++ src/main/java/com/example/springLoan/repository/DataTypeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DataTypeRepository extends JpaRepository<DataType, Integer> {
+public interface DataTypeRepository extends JpaRepository<DataType, Long> {
 }
