+++ src/main/java/com/example/springLoan/repository/SettingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SettingRepository extends JpaRepository<Setting, Integer> {
+public interface SettingRepository extends JpaRepository<Setting, Long> {
 }
