+++ src/main/java/com/example/springLoan/service/CustomerService.java
@@ -5,5 +5,5 @@ import com.example.springLoan.model.Customer;
 import java.util.Optional;
 
 public interface CustomerService {
-    Optional<Customer> findById(Integer id);
+    Optional<Customer> findById(Long id);
 }
