+++ src/main/java/com/example/springLoan/service/ProductTypeService.java
@@ -5,5 +5,5 @@ import com.example.springLoan.model.ProductType;
 import java.util.Optional;
 
 public interface ProductTypeService {
-    Optional<ProductType> findById(Integer id);
+    Optional<ProductType> findById(Long id);
 }
